Moved navigation to a Wordpress Menu/Navigation
Enabled Menus in Wordpress, by ‘add_theme_support’ in functions.php. So
now Menus appear under Appearance (in wordpress).

Created a custom header menu - so with the overflow list and home
added. This needed a custom ’walker’, which is a nav/menu
generator. - Note this applies to every list atm.

Registered the custom menu ‘header-menu’. So it appears as a menu.

Removed old way of creating the menu.