[WIP] Used the API to print the schedule
Print the day of the week to confirm shows on that day.
Add a schedule-live class if the show is currently on.
Create a permalink to the show page based on the give slug.
Print the name of the show.
Need to get the host of the show.
Still need to calculate the length of a show, @jamesturner we might
need a length variable in the show object.