Added our typography styles to the WP post editor
The editor when writing a Wordpress post now looks like how it will
output. e.g. h1 will look like the output h1

The editor toolbar items are defined in functions.php.

The editor stylings are defined in editor-style.css, we import our
output typography styles into the stylesheet, so the input looks the
same as the output.